// ts/types.ts
export type HexString = `0x${string}`
// ...
export type TToken = {
  moonpay_symbol?: string
  token_symbol: string;
  coin_symbol: string;
  name: string;
  address: `0x${string}`;
  default_decimals: number;
  coingecko_token_id: string;
  coin_main_hex_color: string;
  about: string;
  website: string;
  whitepaper: string;
  content: string;
  twitter: string;
  reddit: string;
  telegram: string;
  blockchain_explorer: string;
  coingecko: string;
  messari: string;
  erc20_address?: string;
  logoURI: string;
  priceUSD: number;
  chainId?: number;
  amount: string;
};
// ...
export type TTokenLifi = {
  address: string;
  chainId: number;
  coinKey: string;
  decimals: number;
  logoURI: string;
  name: string;
  priceUSD: string;
  symbol: string;
};

export type TInsurance = {
  feeAmountUsd: string;
  state: string;
};

export type TGasCost = {
  amount: string;
  amountUSD: string;
  estimate: string;
  limit: string;
  price: string;
  token: TTokenLifi;
  type: string;
};

export type TFeeCost = {
  amount: string;
  amountUSD: string;
  description: string;
  included: boolean;
  name: string;
  percentage: string;
  token: TTokenLifi;
};

export type TEstimate = {
  approvalAddress: string;
  executionDuration: number;
  feeCosts: TFeeCost[];
  fromAmount: string;
  fromAmountUSD: string;
  gasCosts: TGasCost[];
  toAmount: string;
  toAmountMin: string;
  toAmountUSD: string;
  tool: string;
};

export type TToolDetails = {
  key: string;
  logoURI: string;
  name: string;
};

export type TAction = {
  fromAmount: string;
  fromChainId: number;
  fromToken: TTokenLifi;
  slippage: number;
  toChainId: number;
  toToken: TTokenLifi;
};

export type TIncludedSteps = {
  action: TAction[];
  integrator: string;
  tool: string;
  toolDetails: TToolDetails;
  type: string;
};

export type TStep = {
  action: TAction;
  estimate: TEstimate;
  id: string;
  includedSteps: TIncludedSteps[];
  integrator: string;
  tool: string;
  toolDetails: TToolDetails;
  type: string;
};

export type TRoutes = {
  containsSwitchChain: boolean;
  fromAmount: string;
  fromAmountUSD: string;
  fromChainId: string;
  fromToken: TTokenLifi;
  gasCostUSD: string;
  id: string;
  insurance: TInsurance;
  steps: TStep[];
  tags: string[];
  toAmount: string;
  toAmountMin: string;
  toAmountUSD: string;
  toChainId: number;
  toToken: TTokenLifi;
};